added alert

+++ src/lib/components/ListView.js
@@ -1,13 +1,25 @@
 import React from 'react';
-
+import DeleteConfirm from './DeleteConfirm';
 
 
 const ListView = ({ onEdit,dataList,deleteItem,pageName="Your Page Name" }) => {
 
 
-    const handleDelete = (id) => {
-      deleteItem(id)
-      };
+    const [isShowDeleteModel, setIsShowDeleteModel] = React.useState(false);
+    const [deleteID, setDeleteID] = React.useState(0);
+
+      const onDeleteConfirm = (id) => {
+        if(id){
+            deleteItem(id)
+        }
+      
+        setIsShowDeleteModel(false);
+      }
+      const handleDelete = (id) => {
+        debugger;
+        setDeleteID(id);
+        setIsShowDeleteModel(true);
+      }
 
   return (
     <div className='App'>
@@ -47,6 +59,9 @@ const ListView = ({ onEdit,dataList,deleteItem,pageName="Your Page Name" }) => {
         </tbody>
       </table>
     </div>
+
+   
+    <DeleteConfirm onDelete={onDeleteConfirm} deleteID={deleteID} showModel={isShowDeleteModel} />
     </div>
   );
 };
